Declare Command and CommandContext globally
