+++ src/commands/help.test.ts
@@ -1,4 +1,3 @@
-import type { CommandContext } from '../CommandContext';
 import { help } from './help';
 
 describe('help', () => {
