+++ src/commands/help.ts
@@ -1,5 +1,3 @@
-import type { GlobalCommand } from '../Command';
-
 export const help: GlobalCommand = {
 	name: 'help',
 	description: 'Prints the list of commands',
