+++ src/helpers/guards/isGuildedCommandContext.test.ts
@@ -1,4 +1,3 @@
-import type { CommandContext, GuildedCommandContext } from '../../CommandContext';
 import type { Guild, GuildMember } from 'discord.js';
 import { ChannelType } from 'discord.js';
 import { isGuildedCommandContext } from './isGuildedCommandContext';
